// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/project-management.model.ts
export class ProjectManagementModel {
    uuid: string;
    projectConceptUuid: any;
    projectConceptId: number;
    nameOfThePost: string;
    quantity: string;
    qualification: string;
    modeOfRecruitment: string;
    scale_amount: string;
    payGrade: string;
    responsibility: string;
    remarks: string;
}